Move away from using paris for logging
Plain log with env_logger produces better output than paris so use that
combination. The main reason is that log + env_logger includes module
names in the log messages that helps a lot when trying to understand
where different messages are coming from. Could be that paris also
supports this, but since log + env_logger is good enough, move to that
one.